pyproject.toml
[project]
name = "snu-seminar-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4",
    "pillow",
    "pytesseract",
    "pytz",
    "requests",
]
